LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-grid-001-expected.txt


PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-y="100"><span></span></div>
offsetTop expected 100 but got 10
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-y="55"><span></span></div>
offsetTop expected 55 but got 25
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-y="115"><span></span></div>
offsetTop expected 115 but got 10
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<div data-offset-y="90"><span></span></div>
offsetTop expected 90 but got 10
PASS .target > * 12
